% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{filtre_genre}
\alias{filtre_genre}
\title{Filtrer les données Netflix par genre}
\usage{
filtre_genre(data, gender)
}
\arguments{
\item{data}{Un data frame.}

\item{gender}{Le genre à conserver.}
}
\value{
Un data frame filtré.
}
\description{
Filtre le jeu de données selon le genre.
}
