% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{data}
\name{netflix}
\alias{netflix}
\title{Jeu de données Netflix}
\format{
Un data frame avec 1000 lignes et plusieurs variables :
\describe{
\item{age}{Âge de l'utilisateur}
\item{gender}{Genre de l'utilisateur}
\item{country}{Pays}
\item{subscription_type}{Type d'abonnement}
\item{avg_watch_time_minutes}{Temps moyen de visionnage en minutes}
}
}
\usage{
netflix
}
\description{
Sous-ensemble du jeu de données sur le comportement des utilisateurs Netflix.
}
\keyword{datasets}
