% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{temps_moyen_abonnement_graph}
\alias{temps_moyen_abonnement_graph}
\title{Faire un graphique du temps de visionnage moyen par abonnement}
\usage{
temps_moyen_abonnement_graph(data)
}
\arguments{
\item{data}{Un data frame.}
}
\value{
Un graphique ggplot2.
}
\description{
Crée un graphique ggplot2 à partir du calcul groupé du temps
de visionnage moyen par type d'abonnement.
}
