% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{temps_moyen_age_pays}
\alias{temps_moyen_age_pays}
\title{Calculer un temps de visionnage moyen après filtrage}
\usage{
temps_moyen_age_pays(data, country, min_age)
}
\arguments{
\item{data}{Un data frame.}

\item{country}{Le pays à conserver.}

\item{min_age}{L'âge minimal à conserver.}
}
\value{
Une valeur numérique.
}
\description{
Filtre les données selon un pays et un âge minimal, puis calcule
le temps de visionnage moyen.
}
