% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{temps_moyen_abonnement}
\alias{temps_moyen_abonnement}
\title{Calculer le temps moyen de visionnage par abonnement}
\usage{
temps_moyen_abonnement(data)
}
\arguments{
\item{data}{Un data frame.}
}
\value{
Un data frame.
}
\description{
Calculer le temps moyen de visionnage par abonnement
}
