% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{temps_moyen_abonnement_graph}
\alias{temps_moyen_abonnement_graph}
\title{Représenter le temps moyen de visionnage par abonnement}
\usage{
temps_moyen_abonnement_graph(data)
}
\arguments{
\item{data}{Un data frame.}
}
\value{
Un graphique ggplot2.
}
\description{
Représenter le temps moyen de visionnage par abonnement
}
